+++ test/e2e/configuration.dart
@@ -34,6 +34,7 @@ import 'mock/graphql.dart';
 import 'mock/platform_utils.dart';
 import 'parameters/attachment.dart';
 import 'parameters/availability_status.dart';
+import 'parameters/credentials.dart';
 import 'parameters/download_status.dart';
 import 'parameters/enabled_status.dart';
 import 'parameters/exception.dart';
@@ -105,6 +106,7 @@ import 'steps/sees_muted_chat.dart';
 import 'steps/select_text.dart';
 import 'steps/sends_attachment.dart';
 import 'steps/sends_message.dart';
+import 'steps/set_login.dart';
 import 'steps/tap_chat.dart';
 import 'steps/tap_chat_in_search_view.dart';
 import 'steps/tap_contact.dart';
@@ -157,6 +159,8 @@ final FlutterTestConfiguration gherkinTestConfiguration =
         dragContactDown,
         fillField,
         fillFieldN,
+        fillFieldWithMyCredential,
+        fillFieldWithUserCredential,
         goToUserPage,
         hasContacts,
         hasDialogWithMe,
@@ -182,6 +186,7 @@ final FlutterTestConfiguration gherkinTestConfiguration =
         longPressContact,
         longPressMessageByAttachment,
         longPressMessageByText,
+        longPressMonolog,
         longPressWidget,
         monologAvailability,
         noInternetConnection,
@@ -215,6 +220,7 @@ final FlutterTestConfiguration gherkinTestConfiguration =
         seeDraftInDialog,
         seeFavoriteChatPosition,
         seeMonologAsFavorite,
+        seeMonologInSearchResults,
         seeNoChatsDismissed,
         seeNoContactsDismissed,
         seesAs,
@@ -225,6 +231,8 @@ final FlutterTestConfiguration gherkinTestConfiguration =
         sendsAttachmentToMe,
         sendsMessageToMe,
         sendsMessageWithException,
+        setLogin,
+        setMyLogin,
         signInAs,
         tapChat,
         tapContact,
@@ -271,6 +279,7 @@ final FlutterTestConfiguration gherkinTestConfiguration =
       ..customStepParameterDefinitions = [
         AttachmentTypeParameter(),
         AvailabilityStatusParameter(),
+        CredentialsParameter(),
         DownloadStatusParameter(),
         EnabledParameter(),
         ExceptionParameter(),